% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{re2_get_options}
\alias{re2_get_options}
\title{Retrieve options}
\usage{
re2_get_options(re2ptr)
}
\arguments{
\item{re2ptr}{The value obtained from call to \code{\link{re2_regexp}}.}
}
\value{
A list of options and their values.
}
\description{
\code{re2_get_options} returns a list of all options from a
  RE2 object (internal representation of compiled regexp).
}
\seealso{
\code{\link{re2_regexp}}.
}
